requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
openai>=0.27.0
python-dotenv>=1.0.0
slowapi>=0.1.4
prometheus-fastapi-instrumentator>=5.12.0
python-jose[cryptography]>=3.3.0
httpx>=0.24.0
sqlalchemy>=1.4
passlib[bcrypt]>=1.7.4
